actually fixed the coordinates this time just need to get out the vetexes somehow :(

+++ src/resources/qhull-2020.2/src/libqhull/global.c
@@ -61,7 +61,7 @@ void setNOerrexit(){
   qh NOerrexit=False;
 }
 
-void coordTset(coordT *ary, int index, float value){
+void coordTset(coordT *ary, int index, double value){
   ary[index] = value;
 }
